refactor: Helper functions and FilterSection component of the pages like admissions, events and notifications are unified
