+++ app/[locale]/academics/admission/page.tsx
@@ -5,6 +5,7 @@ import { arrayOverlaps, desc, eq, inArray } from 'drizzle-orm';
 import { getTranslations } from '~/i18n/translations';
 import { db } from '~/server/db';
 import { cn } from '~/lib/utils';
+import { buildHref, parseDate, toArray } from '~/lib/helpers';
 import ImageHeader from '~/components/image-header';
 import { Button } from '~/components/buttons';
 import { ScrollArea } from '~/components/ui';
@@ -16,6 +17,7 @@ import {
   SearchInput,
 } from '~/components/inputs';
 import { MobileFilters } from '~/components/mobile-filters';
+import { FilterSection } from '~/components/filter-section';
 
 import { NotificationsList } from '../../notifications/NotificationsList';
 
@@ -273,55 +275,4 @@ export default async function AdmissionPage({
       </section>
     </>
   );
-}
-
-/* ---------------------- Filters Components ---------------------- */
-
-function FilterSection({
-  label,
-  children,
-}: {
-  label: string;
-  children: React.ReactNode;
-}) {
-  return (
-    <section className="rounded border border-primary-100 bg-neutral-50 p-4">
-      <div className="flex items-start justify-between">
-        <h3 className="text-xl font-bold text-primary-300">{label}</h3>
-      </div>
-      {children}
-    </section>
-  );
-}
-
-/* ---------------------- Helpers ---------------------- */
-function toArray(v: string | string[] | undefined): string[] {
-  return Array.isArray(v) ? v : v ? [v] : [];
-}
-
-function parseDate(d?: string) {
-  if (!d) return undefined;
-  const date = new Date(d);
-  return isNaN(date.getTime()) ? undefined : date;
-}
-
-function buildHref(locale: string, updates: Record<string, unknown>): string {
-  const params = new URLSearchParams();
-
-  Object.entries(updates).forEach(([k, v]) => {
-    if (v === undefined || (Array.isArray(v) && v.length === 0)) {
-      return;
-    }
-
-    if (Array.isArray(v)) {
-      v.forEach((item) => {
-        if (item) params.append(k, String(item));
-      });
-    } else {
-      params.set(k, String(v));
-    }
-  });
-
-  const qs = params.toString();
-  return `/${locale}/academics/admission${qs ? `?${qs}` : ''}`;
 }
