+++ components/filter-section.tsx
@@ -1,0 +1,30 @@
+import Link from 'next/link';
+import React from 'react';
+
+export function FilterSection({
+  label,
+  children,
+  viewAllHref,
+}: {
+  label: string;
+  children: React.ReactNode;
+  viewAllHref?: string;
+  locale?: string;
+}) {
+  return (
+    <section className="rounded border border-primary-100 bg-neutral-50 p-4">
+      <div className="flex items-start justify-between">
+        <h3 className="text-xl font-bold text-primary-300">{label}</h3>
+        {viewAllHref && (
+          <Link
+            href={viewAllHref}
+            className="text-xs font-medium text-primary-700"
+          >
+            View All
+          </Link>
+        )}
+      </div>
+      {children}
+    </section>
+  );
+}
